增加 单位名称

+++ mqc/models/mqc_blood.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from openerp import models, fields, api
-from openerp.exceptions import ValidationError
+
+
 class BloodClinic(models.Model):
     _name = "mqc.blood.clinic"
     _description = u'全院临床用血情况'
@@ -23,29 +24,29 @@ class BloodClinic(models.Model):
     use_blood_case = fields.Integer(u'用血总人数',)
     opr_num  = fields.Integer(u'手术台数',)
     opr_use_blood = fields.Integer( u'手术用血人数',)
-    whole_blood = fields.Integer( u'全血',)
+    whole_blood = fields.Float(u'全血', )
     short_sign1 = fields.Char(u'全血是否短缺',)
-    rbc = fields.Integer(u'红细胞',)
+    rbc = fields.Float(u'红细胞', )
     short_sign2 = fields.Char(u'红细胞是否短缺',)
-    plasma = fields.Integer(u'血浆',)
+    plasma = fields.Float(u'血浆', )
     short_sign3 = fields.Char(u'血浆是否短缺',)
-    platelet = fields.Integer( u'血小板',)
+    platelet = fields.Float(u'血小板', )
     short_sign4 = fields.Char(u'血小板是否短缺',)
-    cryoprecipitate = fields.Integer( u'冷沉淀',)
+    cryoprecipitate = fields.Float(u'冷沉淀', )
     short_sign5 = fields.Char(u'冷沉淀是否短缺',)
 
-    total_usage = fields.Integer(u'本月用血量合计',)
+    total_usage = fields.Float(u'本月用血量合计', )
     self_storage = fields.Integer( u'储存式',)
     self_storage_case = fields.Integer(u'储存式人数',)
     self_recycle = fields.Float(u'回收式',)
     self_recycle_case = fields.Integer(  u'回收式人数',)
-    dilution = fields.Integer(u'稀释式',)
+    dilution = fields.Float(u'稀释式', )
     dilution_case = fields.Integer(u'稀释式人数',)
-    total_self_usage = fields.Integer( u'自体输血量',)
-    total_self_case = fields.Integer( u'自体输血数',)
+    total_self_usage = fields.Float(u'自体输血量', )
+    total_self_case = fields.Float(u'自体输血数', )
     component_rate = fields.Float(  u'成分输血率(%)',)
     self_rate = fields.Float( u'自体输血率(%)',)
-    avg_usage= fields.Integer(u'住（出）院患者人均用血量',)
+    avg_usage = fields.Float(u'住（出）院患者人均用血量', )
 
 
     # @api.model
